Move wrappers to their own module
In this commit, the module is kept private and its items
re-exported from the crate root. We name it `wrappers`,
but that's a bad (if accurate) name. We can bikeshed the
real name if/when we decide to make it pub.

Makes progress on #252
